+++ examples/simple_issue_nfe.py
@@ -30,7 +30,7 @@ def main():
             description="Test product",
             amount=100.00,
             address=Address(state="SP"),
-            document=Product(ncm="84713012", cfop="5102"),
+            extra=Product(ncm="84713012", cfop="5102"),
         )
     except ConnectionFailedError:
         print("invoice-api is not running on localhost:8000")